requirements.txt
python-dotenv==1.1.1
colorama==0.4.6
google-generativeai==0.8.5
requests==2.32.5
flask==3.0.0
flask-cors==4.0.0
